requirements.txt
APScheduler
Flask
groq
nemo
nltk
omegaconf
pandas
protobuf
psycopg2_binary
python-dotenv
SQLAlchemy
streamlit
torchaudio
tqdm
transformers